+++ sqlalchemy_cockroachdb/psycopg2.py
@@ -1,7 +1,8 @@
 from sqlalchemy.dialects.postgresql.psycopg2 import PGDialect_psycopg2
 from .base import CockroachDBDialect
 from .ddl_compiler import CockroachDDLCompiler
-from .stmt_compiler import CockroachCompiler, CockroachIdentifierPreparer
+from .stmt_compiler import CockroachCompiler
+from .stmt_compiler import CockroachIdentifierPreparer
 
 
 class CockroachDBDialect_psycopg2(PGDialect_psycopg2, CockroachDBDialect):